+++ CHANGELOG.md
@@ -83,6 +83,7 @@ same type.
 - [#364](https://github.com/influxdb/telegraf/pull/364): Support InfluxDB UDP output.
 - [#370](https://github.com/influxdb/telegraf/pull/370): Support specifying multiple outputs, as lists.
 - [#372](https://github.com/influxdb/telegraf/pull/372): Remove gosigar and update go-dockerclient for FreeBSD support. Thanks @MerlinDMC!
+- [#382](https://github.com/influxdb/telegraf/pull/382): Add system wide network protocol stats to `net` plugin.
 
 ### Bugfixes
 - [#331](https://github.com/influxdb/telegraf/pull/331): Dont overwrite host tag in redis plugin.